requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-Mail==0.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.1.2
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.4
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
